:recycle: Move session manager interface stuff to its own module.
