+++ session_manager/views/partials.py
@@ -4,7 +4,7 @@ from django.contrib.auth.decorators import login_required
 from django.http import HttpRequest, HttpResponse
 from django.shortcuts import render
 
-from interfaces import controller_interface as ci
+from interfaces import session_manager_interface as smi
 
 from .. import tables
 
@@ -14,7 +14,7 @@ def active_sessions_table_view(request: HttpRequest) -> HttpResponse:
     """View that renders the active sessions table."""
     return render(
         request=request,
-        context=dict(table=tables.ActiveSessions(ci.get_sessions())),
+        context=dict(table=tables.ActiveSessions(smi.get_sessions())),
         template_name="session_manager/partials/table_partial.html",
     )
 
@@ -24,6 +24,6 @@ def available_configs_table_view(request: HttpRequest) -> HttpResponse:
     """View that renders the available configs table."""
     return render(
         request=request,
-        context=dict(table=tables.AvailableConfigs(ci.get_configs())),
+        context=dict(table=tables.AvailableConfigs(smi.get_configs())),
         template_name="session_manager/partials/table_partial.html",
     )
